requirements.txt
eventlet==0.30.2
python-socketio
asyncio
gunicorn
pandas
